requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
colorama==0.4.6
idna==3.4
python-dotenv==0.21.0
requests==2.28.1
shodan==1.28.0
urllib3==1.26.12
XlsxWriter==3.0.3
